Enforce single class per file
