+++ rules.ruleset
@@ -130,8 +130,8 @@
     <Rule Id="SA1314" Action="None" />
     <Rule Id="SA1400" Action="Warning" />
     <Rule Id="SA1401" Action="None" />
-    <Rule Id="SA1402" Action="None" />
-    <Rule Id="SA1403" Action="None" />
+    <Rule Id="SA1402" Action="Warning" />
+    <Rule Id="SA1403" Action="Warning" />
     <Rule Id="SA1404" Action="Warning" />
     <Rule Id="SA1405" Action="None" />
     <Rule Id="SA1406" Action="None" />
