+++ src/DocumentFormat.OpenXml/ObsoleteAttributeMessages.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace DocumentFormat.OpenXml.Packaging
+namespace DocumentFormat.OpenXml
 {
     internal static class ObsoleteAttributeMessages
     {
