+++ src/DocumentFormat.OpenXml/Validation/Schema/ParticleValidator.cs
@@ -1,16 +1,17 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using DocumentFormat.OpenXml.Validation;
 using System;
 using System.Diagnostics;
+
+#pragma warning disable SA1402 // File may only contain a single type
 
 namespace DocumentFormat.OpenXml.Validation.Schema
 {
 #if DEBUG
-    /// <summary>
-    /// InstanceCounter for performance investigation.
-    /// </summary>
+                              /// <summary>
+                              /// InstanceCounter for performance investigation.
+                              /// </summary>
     public static class InstanceCounter
     {
         /// <summary>
