+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/AppendCommentExIDs.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Element CommentEx in paraId attribute
+    /// </summary>
+    public static class AppendCommentExIDs
+    {
+        public static string AppendCommentID1 { get { return "11111111"; } }
+
+        public static string AppendCommentID2 { get { return "22222222"; } }
+    }
+}
