+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/CommentAuthors.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Element Comment attribute "Author"
+    /// </summary>
+    public static class CommentAuthors
+    {
+        public static string Author1 { get { return "Masaki Tamura (Pasona Tech)"; } }
+
+        public static string Author2 { get { return "Dan Ito"; } }
+    }
+}
