+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/CommentBodyStrings.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Word document body text
+    /// </summary>
+    public static class CommentBodyStrings
+    {
+        public static string Comment1 { get { return "Test1"; } }
+
+        public static string Comment2 { get { return "Test2"; } }
+    }
+}
