+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/CommentIDs.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Element Comment attribute "ID"
+    /// </summary>
+    public static class CommentIDs
+    {
+        public static string CommentID1 { get { return "1"; } }
+
+        public static string CommentID2 { get { return "2"; } }
+
+        public static string CommentID3 { get { return "3"; } }
+
+        public static string CommentID4 { get { return "4"; } }
+
+        public static string CommentID5 { get { return "5"; } }
+    }
+}
