+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/CommentInitials.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Element Comment attribute "Initial"
+    /// </summary>
+    public static class CommentInitials
+    {
+        public static string Initial1 { get { return "M.T"; } }
+
+        public static string Initial2 { get { return "D.I"; } }
+    }
+}
