+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/CommentEx/CommentStrings.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.CommentEx
+{
+    /// <summary>
+    /// Comment text
+    /// </summary>
+    public static class CommentStrings
+    {
+        public static string CommentString1 { get { return "Comment1"; } }
+
+        public static string CommentString2 { get { return "Comment2"; } }
+
+        public static string CommentString3 { get { return "Comment3"; } }
+
+        public static string CommentChangeString1 { get { return "Changed comment1"; } }
+
+        public static string CommentAppendString1 { get { return "Appended comment1"; } }
+
+        public static string CommentAppendString2 { get { return "Appended comment2"; } }
+    }
+}
