+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/ContentControl/RunStyleValues.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.ContentControl.ConstantStrings
+{
+    public static class RunStyleValues
+    {
+        public const string Placeholder = "PlaceholderText";
+    }
+}
