+++ test/DocumentFormat.OpenXml.Tests/ConformanceTest/ContentControl/TestStrings.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.ContentControl.ConstantStrings
+{
+    public static class TestStrings
+    {
+        public const string ContentControlString1 = "Click here to enter text.";
+        public const string ContentControlString2 = "Test string";
+        public const string ContentControlCellString1 = "Repeating Section,Row1,Cell1";
+        public const string ContentControlCellString2 = "Repeating Section,Row1,Cell2";
+        public const string ContentControlCellString3 = "Repeating Section,Row2,Cell1";
+        public const string ContentControlCellString4 = "Repeating Section,Row2,Cell2";
+    }
+}
